getmoney/requirements.txt
amqp==2.5.2
asgiref==3.2.5
billiard==3.6.3.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.0.0
django-filters==0.2.1
django-rest-auth==0.9.5
django-sendsms==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drfaddons==0.1.3
flower==0.9.4
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
oauthlib==3.1.0
pkg-resources==0.0.0
pycrypto==2.6.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
